+++ tests/testthat/test-wrangle-helpers.R
@@ -6,3 +6,11 @@ test_that("lines_to_matrix works", {
 							 				 "#", "#", ".", "."
 							 				 ), byrow = TRUE, nrow = 3))
 })
+
+# write test for split_at_blanks
+test_that("split_at_blanks works", {
+	x <- c("123", "1234", "", "56", "567", "", "89")
+	expect_equal(split_at_blanks(x), list(`1` = c("123", "1234"), `2` = c("56", "567"), `3` = "89"))
+	expect_equal(split_at_blanks(x, "numeric"), list(`1` = c(123, 1234), `2` = c(56, 567), `3` = 89))
+
+})