+++ NEWS.md
@@ -5,4 +5,5 @@
 * Add user agent to `aoc_get_input()`
 * Add `extract_numbers()` to extract all numbers from a string
 * Add `GCD()` and `LCM()` for finding greatest common divisor and lowest common multiple
-* Add `lines_to_matrix()`
+* Add `lines_to_matrix()` to convert a vector of strings to a matrix
+* Add `split_at_blanks()` to create list from input, with one element per group originally separated by blank lines